[Modern Media Controls] CSS "cursor" property shoud be respected in media controls shadow root https://bugs.webkit.org/show_bug.cgi?id=213295 <rdar://problem/61911638>
Reviewed by Timothy Hatcher.

Source/WebCore:

Allow the "cursor" property to be inherited in the media controls shadow root, but still overriden
for interactive objects in the media controls as well as placard text.

Test: media/modern-media-controls/css/cursor.html

* Modules/modern-media-controls/controls/controls-bar.css:
(.controls-bar):
* Modules/modern-media-controls/controls/media-controls.css:
(.media-controls-container):
(.media-controls):
* Modules/modern-media-controls/controls/placard.css:
(.placard .title,):

LayoutTests:

Add a test that checks that setting the "cursor" property on a <video> element also applies to the
media controls in the element's shadow root.

* media/modern-media-controls/css/cursor-expected.txt: Added.
* media/modern-media-controls/css/cursor.html: Added.

Identifier: 226084@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263153 268f45cc-cd09-0410-ab3c-d52691b4dbfc