Disable the JS JIT when running in a translated process https://bugs.webkit.org/show_bug.cgi?id=213478
Reviewed by Saam Barati.

Source/JavaScriptCore:

* runtime/Options.cpp:
(JSC::Options::recomputeDependentOptions):
Based on our performance experiements, disable the JavaScript JIT
(but not the regular expression, DOM, or Wasm JIT) when running
in a translated process.

Source/WTF:

* wtf/PlatformHave.h:
* wtf/TranslatedProcess.h: Added.
* WTF.xcodeproj/project.pbxproj:
* wtf/CMakeLists.txt:
Add a helper function that can be used to determine that the current process
is being translated; currently the only case is an x86_64 process running on arm64e.


Identifier: 226271@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263365 268f45cc-cd09-0410-ab3c-d52691b4dbfc