Expose already-available WKNavigationAction.request.HTTPBody through the API https://bugs.webkit.org/show_bug.cgi?id=179077
Patch by Alex Christensen <achristensen@webkit.org> on 2020-06-23
Reviewed by Brady Eidson.

Source/WebKit:

When this bug was written, we needed to serialize the HTTP body to the UI process to expose it through the API.
The controversial part of this was done in r237639, so this just exposes the data that is already available.

Verified by an API test.

* UIProcess/API/Cocoa/WKNavigationAction.mm:
(-[WKNavigationAction request]):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/Navigation.mm:
(TEST):

Identifier: 226315@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263411 268f45cc-cd09-0410-ab3c-d52691b4dbfc